+++ tests/quotaSinglePool.test.js
@@ -73,7 +73,7 @@ function resetMocks() {
     global.isCalculating = false;
 }
 
-let appJsCode = fs.readFileSync('app.js', 'utf8');
+let appJsCode = fs.readFileSync('translations.js', 'utf8') + '\n' + fs.readFileSync('app.js', 'utf8');
 
 // Strip 'let' so app.js uses our global variables
 appJsCode = appJsCode.replace(/let doctors = /g, 'global.doctors = ');