+++ tests/scoring.test.js
@@ -12,7 +12,7 @@ global.window = {
 };
 global.localStorage = { getItem: () => 'en', setItem: () => {} };
 
-let appJsCode = fs.readFileSync('./app.js', 'utf8');
+let appJsCode = fs.readFileSync('./translations.js', 'utf8') + '\n' + fs.readFileSync('./app.js', 'utf8');
 
 // Strip 'let' so app.js uses our global variables
 appJsCode = appJsCode.replace(/let doctors = /g, 'global.doctors = ');