+++ tests/solver.test.js
@@ -69,7 +69,7 @@ function resetMocks() {
 }
 
 // Load app.js code
-let appJsCode = fs.readFileSync('./app.js', 'utf8');
+let appJsCode = fs.readFileSync('./translations.js', 'utf8') + '\n' + fs.readFileSync('./app.js', 'utf8');
 
 eval(appJsCode);