+++ tests/uiVisibility.test.js
@@ -51,7 +51,7 @@ global.isCustomDateRange = false;
 let toastMessages = [];
 
 // Load app.js code
-let appJsCode = fs.readFileSync('app.js', 'utf8');
+let appJsCode = fs.readFileSync('translations.js', 'utf8') + '\n' + fs.readFileSync('app.js', 'utf8');
 
 // Strip 'let' so app.js uses our global variables
 appJsCode = appJsCode.replace(/let doctors = /g, 'global.doctors = ');